ai-blog-agent-template/requirements.txt
langgraph>=0.2.0
langchain-core>=0.2.0
langchain-openai>=0.1.0
python-dotenv>=1.0.0
pydantic>=2.0.0
streamlit>=1.30.0
langchain-community>=0.2.0
# Optional for image generation:
# google-genai>=0.3.0
